Refactor keyvault provider to avoid circular imports
