+++ takeoff/azure/credentials/active_directory_user.py
@@ -1,6 +1,6 @@
 from msrestazure.azure_active_directory import UserPassCredentials
 
-from takeoff.azure.credentials.keyvault_credentials_provider import KeyVaultCredentialsMixin
+from takeoff.azure.credentials.providers.keyvault_credentials_mixin import KeyVaultCredentialsMixin
 from takeoff.util import current_filename
