+++ takeoff/azure/credentials/artifact_store.py
@@ -1,6 +1,6 @@
 from twine.settings import Settings
 
-from takeoff.azure.credentials.keyvault_credentials_provider import KeyVaultCredentialsMixin
+from takeoff.azure.credentials.providers.keyvault_credentials_mixin import KeyVaultCredentialsMixin
 from takeoff.util import current_filename
