+++ takeoff/azure/credentials/providers/keyvault_credentials_mixin.py
@@ -1,12 +1,10 @@
 import re
 from dataclasses import dataclass
-from typing import List, Dict, Optional, Union, Tuple
+from typing import List, Dict, Optional
 
 from azure.keyvault import KeyVaultClient as AzureKeyVaultClient
 from azure.keyvault.models import SecretBundle
 
-from takeoff.azure.credentials.keyvault import KeyVaultClient
-from takeoff.credentials.credential_provider import BaseProvider
 from takeoff.credentials.secret import Secret
 from takeoff.util import get_matching_group, has_prefix_match, inverse_dictionary
 
@@ -140,14 +138,3 @@ class KeyVaultCredentialsMixin(object):
         ]
 
         return app_secrets
-
-
-class AzureKeyVaultProvider(BaseProvider, KeyVaultCredentialsMixin):
-    def __init__(self, config, app_version):
-        super().__init__(config, app_version)
-        self.vault_name, self.vault_client = KeyVaultClient.vault_and_client(self.config, self.env)
-
-    def get_credentials(self, lookup: Union[str, Dict[str, str], Tuple[str, str]]):
-        if not isinstance(lookup, str):
-            raise ValueError("Please provide a string")
-        return self._transform_key_to_credential_kwargs(self.config["azure"]["keyvault_keys"][lookup])
