+++ takeoff/azure/credentials/service_principal.py
@@ -1,5 +1,6 @@
 from msrestazure.azure_active_directory import ServicePrincipalCredentials as SpCredentials
 
+from takeoff.azure.credentials.providers.keyvault_credentials_mixin import KeyVaultCredentialsMixin
 from takeoff.credentials.environment_credentials_provider import EnvironmentCredentialsMixin
 from takeoff.util import current_filename
 
@@ -11,4 +12,11 @@ class ServicePrincipalCredentials(EnvironmentCredentialsMixin):
         )
         return SpCredentials(**credential_kwargs)
 
+
+class ServicePrincipalCredentialsFromVault(KeyVaultCredentialsMixin):
+    def credentials(self, config: dict) -> SpCredentials:
+        credential_kwargs = super()._transform_key_to_credential_kwargs(
+            config["azure"]["keyvault_keys"][current_filename(__file__)]
+        )
+        return SpCredentials(**credential_kwargs)
