+++ takeoff/azure/deploy_to_kubernetes.py
@@ -14,9 +14,8 @@ from kubernetes.client import CoreV1Api
 from takeoff.application_version import ApplicationVersion
 from takeoff.azure.credentials.active_directory_user import ActiveDirectoryUserCredentials
 from takeoff.azure.credentials.keyvault import KeyVaultClient
-from takeoff.azure.credentials.keyvault_credentials_provider import KeyVaultCredentialsMixin
-from takeoff.azure.credentials.service_principal import ServicePrincipalCredentials
-from takeoff.azure.credentials.service_principal_from_keyvault import ServicePrincipalCredentialsFromVault
+from takeoff.azure.credentials.providers.keyvault_credentials_mixin import KeyVaultCredentialsMixin
+from takeoff.azure.credentials.service_principal import ServicePrincipalCredentialsFromVault
 from takeoff.azure.credentials.subscription_id import SubscriptionId
 from takeoff.azure.util import get_resource_group_name, get_kubernetes_name
 from takeoff.context import Context, ContextKey
