+++ tests/azure/credentials/test_keyvault_credentials_provider.py
@@ -2,7 +2,7 @@ from unittest import mock
 
 from azure.keyvault.models import SecretBundle
 
-from takeoff.azure.credentials.keyvault_credentials_provider import KeyVaultCredentialsMixin
+from takeoff.azure.credentials.providers.keyvault_credentials_mixin import KeyVaultCredentialsMixin
 
 
 class TestAzureKeyVaultCredentialsMixin(object):
